# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Api Rest Clon de Airbnb",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/users" : {
            "get": {
                "description": "Obtener todos mis usuarios si eres administrador",
                "tags": ["Users", "v1"],
                "responses": {
                    "200": {
                        "description": "Obtenemos exitosamente todos los usuarios"
                    },
                    "400": {
                        "description": "Generamos un error en caso de que la peticion no se resulva"
                    }
                }
            }
        },
...
        "/api/v1/accommodations": {
            "get": {
                "description": "Obtener todos los lugares",
                "tags": ["accommodation", "v1"],
                "responses": {
                    "200": {
                        "description": "Obtenemos exitosamente todos los lugares"
                    },
                    "400": {
                        "description": "Generamos un error en caso de que la peticion no se resulva"
                    }
                }
            }
        },
...
    }
}